cli: Run startup tasks in the Dangerzone CLI
Now that Dangerzone is based on the embedded Podman runtime, the
Dangerzone CLI needs to initialize and start the Podman machine before
starting any conversion. Even more importantly, it should check for
updates and install the proper container image.

Those things are defined as startup tasks, and in this commit we let the
Dangerzone CLI take advantage of them, and run them synchronously when
it starts.

Note that starting and stopping Podman machines can take some time,
which is especially pronounced in CI tests. For these reason, we add a
flag that allows users to let the Podman machine linger, if they want to
fire up multiple conversion tasks in a row.